+++ src/pragzip/huffman/HuffmanCodingSymbolsPerLength.hpp
@@ -12,6 +12,17 @@
 
 namespace pragzip
 {
+/**
+ * This is an iterative improvement over HuffmanCodingLinearSearch.
+ * - During initialization, it stores all symbols (for each code) sorted by length in an array and also stores
+ *   offsets to jump to subarrays of symbols with a given code length. The subarray size is given by the next offset.
+ *   This avoids going over all elements all the time and also already implements usage of maximum-sized and
+ *   manually managed memory chunks by using std::array to avoid heap allocations.
+ * - During decoding, it reads the bits one by one and for each intermediary, checks whether
+ *   there is a matching code, which is calcualted ad-hoc from m_minimumCodeValuesPerLevel.
+ *   If the code matches, it gets the symbol for current code length from the corresponding subarray.
+ * Note that reading the bits one by one is also necessary to reverse the codes.
+ */
 template<typename HuffmanCode,
          uint8_t  MAX_CODE_LENGTH,
          typename Symbol,
@@ -36,7 +47,7 @@ protected:
             m_offsets[bitLength - this->m_minCodeLength] = sum;
             sum += bitLengthFrequencies[bitLength];
         }
-        m_offsets[this->m_maxCodeLength - this->m_minCodeLength + 1] = sum;
+        m_offsets[this->m_maxCodeLength - this->m_minCodeLength + 1] = static_cast<uint16_t>( sum );
 
         /* The codeLengths.size() check above should implicitly check this already. */
         assert( sum <= m_symbolsPerLength.size() && "Specified max symbol range exceeded!" );
@@ -87,7 +98,7 @@ public:
          * would be inversed. @todo Reverse the Huffman codes and prepend bits instead of appending, so that this
          * first step can be conflated and still have the correct order for comparison! */
         for ( BitCount i = 0; i < this->m_minCodeLength; ++i ) {
-            code = ( code << 1 ) | ( bitReader.read<1>() );
+            code = ( code << 1U ) | ( bitReader.read<1>() );
         }
 
         for ( BitCount k = 0; k <= this->m_maxCodeLength - this->m_minCodeLength; ++k ) {
